requirements.txt
pymongo>=4.6.3
psycopg2-binary>=2.9.9
python-dateutil>=2.9.0.post0
tenacity>=8.0
python-dotenv>=1.0